completed b1 and b2
- completed b1 and b2
- let me know if anything needs changes or edits.
